+++ circle.yml
@@ -1,0 +1,12 @@
+machine:
+ ruby:
+   version: 2.4.0
+
+dependencies:
+  override:
+    - bundle install --retry=3
+
+test:
+  post:
+    - bundle exec rspec
+    - bundle exec rubocop -R --format simple
