--- lib/wor/push-notifications/aws/version.rb
@@ -1,7 +1,0 @@
-module Wor
-  module PushNotifications
-    module Aws
-      VERSION = '0.1.0'.freeze
-    end
-  end
-end
