+++ lib/wor/push/notifications/aws/version.rb
@@ -1,0 +1,9 @@
+module Wor
+  module Push
+    module Notifications
+      module Aws
+        VERSION = '0.1.0'.freeze
+      end
+    end
+  end
+end
