+++ spec/spec_helper.rb
@@ -1,7 +1,7 @@
 require 'simplecov'
 SimpleCov.start
-$LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
-require "wor/push-notifications/aws"
+$LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
+require 'wor/push/notifications/aws'
 require 'rails'
 
 require 'byebug'
