+++ spec/wor/push/notifications/aws_spec.rb
@@ -1,8 +1,8 @@
-require "spec_helper"
+require 'spec_helper'
 
-describe Wor::PushNotifications::Aws do
-  it "has a version number" do
-    expect(Wor::PushNotifications::Aws::VERSION).not_to be nil
+describe Wor::Push::Notifications::Aws do
+  it 'has a version number' do
+    expect(Wor::Push::Notifications::Aws::VERSION).not_to be nil
   end
 
   describe '.config' do
@@ -88,7 +88,7 @@ describe Wor::PushNotifications::Aws do
 
         it 'raises ArgumentError' do
           expect { wrong_config }.to raise_error(ArgumentError,
-                                                'Argument must not be an empty string')
+                                                 'Argument must not be an empty string')
         end
       end
     end
@@ -128,7 +128,7 @@ describe Wor::PushNotifications::Aws do
           end
         end
         context 'with invalid type (semantically)' do
-          let(:new_device_types) { [:android, :ois] }
+          let(:new_device_types) { %i[android ois] }
           let(:wrong_config) do
             described_class.configure do |config|
               config.device_types = new_device_types
